# By Aleksei Kanash
# Via Aleksei Kanash (4) and Sergei Zdanevich (1)
* commit '8e167c098abd77a4bcfdbe378cdd614c339e58da [formerly 21f21f9aaa9d2ef31a240df8d8be8fb0462874f7]':
  [DF-7335] Color uses C-style array to keep compatibility to external code.
  [DF-7335] Codestyle fix.
  [DF-7335] NMaterial properties is Vectors now.
  [DF-7335] removed COUNT_OF after development merge.
  [DF-7335] Review fixes, correct isdigit usage.
  [DF-7335] Win 10 warnings fix.
  [DF-7335] std::array removed from Vectors.
  [DF-7335] Review fix. Compilation fix.
  [DF-7335] Review fixes.
  [DF-7335] Vegetation system has no COUNT_OF usage. Vectors usage refactor.
  [DF-7335] Vegetation Geometry arrays uses std::array and std::vector.
  [DF-7335] Removed offsetof as non compilable on Mac. Count_Of method usage removed almost everywhere.
  [DF-7335] NMaterial don't uses COUNT_OF.
  [DF-7335] Review fixes, more  nullptr instead of NULL in changed files.
  [DF-7335] Some PVS warnings fixed.
  [DF-7335] Warnings fix for windows x64


Former-commit-id: 3abf1c2588a44e6af383bfee07ae132210bee436

+++ Sources/Internal/Render/Highlevel/Frustum.cpp
@@ -288,7 +288,7 @@ Frustum::eFrustumResult Frustum::Classify(const AABBox3 & box) const
 
 Frustum::eFrustumResult Frustum::Classify(const AABBox3 & box, uint8 &planeMask, uint8 &startId) const
 {
-	const float32* verts[2] = {box.min.data, box.max.data};
+    const float32* verts[2] = { box.min.data, box.max.data };
 	Frustum::eFrustumResult result = EFR_INSIDE;	
 	uint8 k;
 	const Plane *plane, *startPlane;
@@ -312,7 +312,7 @@ Frustum::eFrustumResult Frustum::Classify(const AABBox3 & box, uint8 &planeMask,
 		{	
 			if (plane->DistanceToPoint(verts[currPlaneAccess&1][0], verts[(currPlaneAccess>>1)&1][1], verts[(currPlaneAccess>>2)&1][2]) > 0.0f)
 			{
-				startId = plane-planeArray;
+				startId = static_cast<uint8>(plane-planeArray);
 				return EFR_OUTSIDE;
 			}
 			invPlaneAccess=~currPlaneAccess;
@@ -328,7 +328,7 @@ Frustum::eFrustumResult Frustum::Classify(const AABBox3 & box, uint8 &planeMask,
 
 bool Frustum::IsInside(const AABBox3 & box, uint8 planeMask, uint8& startClippingPlane)const
 {
-	const float32* verts[2] = {box.min.data, box.max.data};
+    const float32* verts[2] = { box.min.data, box.max.data };
 	uint8 k;
 	const Plane *plane, *startPlane;
 	startPlane = planeArray+startClippingPlane;
@@ -345,7 +345,7 @@ bool Frustum::IsInside(const AABBox3 & box, uint8 planeMask, uint8& startClippin
 		{			
 			if (plane->DistanceToPoint(verts[currPlaneAccess&1][0], verts[(currPlaneAccess>>1)&1][1], verts[(currPlaneAccess>>2)&1][2]) > 0.0f)			
 			{
-				startClippingPlane = plane-planeArray;
+				startClippingPlane = static_cast<uint8>(plane-planeArray);
 				return false;
 			}
 		}
